put_data et get_data
