+++ generators/generate_expense_report.py
@@ -8,12 +8,10 @@ import sys, os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from dolibarr_api import *
-from generators.generate_utils import *
 from generators.utils.fill_data import *
+from generators.utils.get_data import *
+from generators.utils.put_data import *
 
-from generators.utils.get_projects_of_contactID import get_projects_of_contactID
-from generators.utils.get_random_project_id import get_random_project_id
-# from generators.utils.put_fk_project import put_fk_project
 
 def generate_expense_report( dateStart, testing = False):
     url = urlBase + "expensereports"
